django/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "тест из обучения"
authors = ["Борис Гельруд <agu-66@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.1.1"
django = "^4.2.7"
django-split-settings = "^1.2.0"
django-debug-toolbar = "^4.2.0"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.31.0"
pydantic-settings = "^2.1.0"
pydantic = "^2.6.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 79
